MR feedback + legislation/applicant type bugfix
