+++ bin/migrate-oats-data/common/__init__.py
@@ -11,3 +11,5 @@ from .oats_temp_tables import *
 from .oats_to_alcs_parcel_enum import *
 from .alcs_init_users import *
 from .oats_decision_enum import *
+from .oats_legislation_to_alcs_applicant_enum import *
+from .oats_to_alcs_nfu_type_enum import *
